Adding addBooks.js module
